# Node48 add: use infinite loop to find an empty children slot